refactor: 섹션 선택 및 좌석 선택 페이지 리팩토링
